.travis.yml:remove linux (checked on master)
I'm not sure exactly what @sexybiggetje had in mind, but for now, the
mac os build can happen on the travis_platforms branch. On the master,
the build takes a few hours to check because it has to wait for the
macos queue. I'd like quicker feedback when pushing to master. I'll try
to keep the two branches in sync though.